www/map-app/* - rewire to supply config via module parameter
Essentially: the old method assumes something can insert a file into
the module source tree containing the confing.  This doesn't work so
well in general, yet alone if we are writing an npm module. So we have
to tackle this in order to allow the module to be configurable in the
normal way.

We tackle it by, for those files defining modules that need config
information, making them accept this config via a parameter to an
initialisation function. This requires chaning a lot of files; we try
to do this in a minimal way which touches only the "wrapper" of the
logic.

There is a knock-on effect where modules which use other modules that
require config, must require them inside their aforementioned
initialisation function.

In some places there already is an `init` function, in which case we
side step this minimally by prepending an underscore to the name.

The end result works insofar as webpack and (I think) the require'ment
goes, but we are still not supplying the get_dataset.php file, so we
are not done yet.